Backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==24.3
python-decouple==3.8
psycopg2-binary==2.9.10
gunicorn==23.0.0
channels==4.1.0
daphne==4.1.2
channels-redis==4.2.0
